requirements.txt
numpy==2.2.2
pika==1.3.2
polars==1.24.0
psycopg2==2.9.10
python-dotenv==1.0.1